backend/requirements.txt
# YAQIZ Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
bcrypt>=4.0.0,<4.1.0

# YOLO & CV
ultralytics>=8.0.26
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Workstation Monitoring (MediaPipe)
mediapipe>=0.10.9
pynput>=1.7.6

# Utilities
python-dotenv>=1.0.0
websockets>=12.0
httpx>=0.27.0
